% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docknitr.R
\name{docker_alias}
\alias{docker_alias}
\title{Create an alias for a Docker Rmarkdown engine}
\usage{
docker_alias(name, ...)
}
\arguments{
\item{name}{The name of a new Docker Rmarkdown engine}

\item{...}{options for the new Docker Rmarkdown engine}
}
\description{
After writing
}
\details{
docker_alias("ubuntu", image="ubuntu:latest", command="bash")

you can use ```{r engine="ubuntu"} to process Rmarkdown chunks through Ubuntu bash using Docker
}
\examples{
docker_alias("ubuntu", image="ubuntu:latest", command="bash")
}
